Modified to be Topic Bot
Removed unnecessary code and changed the bot to search for "005" (the
signal it gets when fully connected) instead of "You have not" (which
also required getting rid of line 8 which prematurely sent the JOIN
request).

Also added code to wait for the edit of the "announcement" folder and
then set the topic to the contents of "announcement.txt" within that
folder.